Move client::bridge into gateway
