+++ examples/e15_simple_dashboard/src/main.rs
@@ -20,10 +20,10 @@ use std::time::Instant;
 use rillrate::prime::table::{Col, Row};
 use rillrate::prime::*;
 use serenity::async_trait;
-use serenity::client::bridge::gateway::{ShardId, ShardManager};
 use serenity::client::{Client, Context, EventHandler};
 use serenity::framework::standard::macros::{command, group, hook};
 use serenity::framework::standard::{CommandResult, StandardFramework};
+use serenity::gateway::{ShardId, ShardManager};
 use serenity::model::prelude::*;
 use serenity::prelude::*;
 use tokio::sync::Mutex;
