+++ src/gateway/bridge/event.rs
