+++ src/gateway/mod.rs
@@ -40,6 +40,7 @@
 //! [`Client::start_shards`]: crate::Client::start_shards
 //! [docs]: https://discordapp.com/developers/docs/topics/gateway#sharding
 
+mod bridge;
 mod error;
 mod shard;
 mod ws;
@@ -50,6 +51,7 @@ use std::fmt;
 use reqwest::IntoUrl;
 use reqwest::Url;
 
+pub use self::bridge::*;
 pub use self::error::Error as GatewayError;
 pub use self::shard::Shard;
 pub use self::ws::WsClient;
